Moved remaining interfaces
