+++ client/src/interfaces/authentication.ts
@@ -20,12 +20,3 @@ export interface JWTReturnAuthToken {
 	jti: string;
 	nbf: number;
 }
-
-export interface NavotarClientFeature {
-	featureName: string;
-	clientFeatureId: number;
-	clientId: number;
-	featureId: number;
-	flag: string | null;
-	value: string | null;
-}
