+++ client/src/interfaces/clients/clientFeatures.ts
@@ -1,0 +1,8 @@
+export interface NavotarClientFeature {
+	featureName: string;
+	clientFeatureId: number;
+	clientId: number;
+	featureId: number;
+	flag: string | null;
+	value: string | null;
+}
