+++ client/src/interfaces/reservations/reservationType.ts
@@ -1,0 +1,12 @@
+export interface ReservationType {
+	typeId: number;
+	typeName: string;
+	calculationType: string;
+	laguages: string | null;
+	isDefault: boolean;
+	abbreviation: string | null;
+	isFront: boolean;
+	isDeleted: boolean;
+	isPossibleDelete: boolean;
+	isAddedByCustomer: boolean;
+}
