+++ client/src/interfaces/vehicles/vehicleStatus.ts
