+++ client/src/interfaces/vehicles/vehicleType.ts
@@ -1,0 +1,4 @@
+export interface VehicleTypeShort {
+	id: number;
+	value: string;
+}
