+++ client/src/redux/slices/appKeyValues.ts
@@ -1,8 +1,10 @@
 import { createSlice, PayloadAction } from "@reduxjs/toolkit";
-import { VehicleStatus } from "../../interfaces/status/vehicleStatus";
-import { AgreementStatus } from "../../interfaces/status/agreementStatus";
-import { ReservationStatus } from "../../interfaces/status/reservationStatus";
-import { ReservationType, AgreementType, VehicleTypeShort } from "../../interfaces/types";
+import { VehicleStatus } from "../../interfaces/vehicles/vehicleStatus";
+import { AgreementStatus } from "../../interfaces/agreements/agreementStatus";
+import { ReservationStatus } from "../../interfaces/reservations/reservationStatus";
+import { ReservationType } from "../../interfaces/reservations/reservationType";
+import { AgreementType } from "../../interfaces/agreements/agreementType";
+import { VehicleTypeShort } from "../../interfaces/vehicles/vehicleType";
 
 import {
 	fetchReservationStatusesThunk,
