+++ client/src/redux/thunks/appKeyValuesThunks.ts
@@ -4,10 +4,12 @@ import appAxiosInstance from "../../api/appAxiosInstance";
 import { Alert } from "rsuite";
 
 import { RootState } from "../store";
-import { VehicleStatus } from "../../interfaces/status/vehicleStatus";
-import { AgreementStatus } from "../../interfaces/status/agreementStatus";
-import { ReservationStatus } from "../../interfaces/status/reservationStatus";
-import { AgreementType, ReservationType, VehicleTypeShort } from "../../interfaces/types";
+import { VehicleStatus } from "../../interfaces/vehicles/vehicleStatus";
+import { AgreementStatus } from "../../interfaces/agreements/agreementStatus";
+import { ReservationStatus } from "../../interfaces/reservations/reservationStatus";
+import { ReservationType } from "../../interfaces/reservations/reservationType";
+import { AgreementType } from "../../interfaces/agreements/agreementType";
+import { VehicleTypeShort } from "../../interfaces/vehicles/vehicleType";
 
 export const fetchReservationStatusesThunk = createAsyncThunk(
 	"appKeyValues/fetchReservationStatuses",
